Split CAPI C++ into domain translation units
