+++ native/lib/CAPI/Beaver.cpp
@@ -4,14 +4,6 @@
 #include "mlir/CAPI/Pass.h"
 #include "mlir/CAPI/Registration.h"
 #include "mlir/CAPI/Rewrite.h"
-#include "mlir-c/Beaver/ActionTracing.h"
-#include "mlir/Debug/BreakpointManagers/TagBreakpointManager.h"
-#include "mlir/Debug/ExecutionContext.h"
-#include "mlir/Dialect/IRDL/IRDLLoading.h"
-#include "mlir/Dialect/Utils/ReshapeOpsUtils.h"
-#include "mlir/IR/Action.h"
-#include "mlir/IR/ExtensibleDialect.h"
-#include "mlir/IR/Unit.h"
 #include "llvm/Config/llvm-config.h"
 #include "llvm/Support/ThreadPool.h"
 #include "llvm/Support/VCSRevision.h"
@@ -312,18 +304,6 @@ MLIR_CAPI_EXPORTED MlirPass beaverPassCreate(
       userData);
 }
 
-MLIR_CAPI_EXPORTED MlirAttribute beaverGetReassociationIndicesForReshape(
-    MlirType sourceType, MlirType targetType) {
-  auto indices = mlir::getReassociationIndicesForReshape(
-      mlir::cast<RankedTensorType>(unwrap(sourceType)),
-      mlir::cast<RankedTensorType>(unwrap(targetType)));
-  OpBuilder b{unwrap(sourceType).getContext()};
-  if (!indices) {
-    return wrap(Attribute{});
-  }
-  return wrap(getReassociationIndicesAttribute(b, *indices));
-}
-
 MLIR_CAPI_EXPORTED void beaverLocationPrint(MlirLocation location,
                                             MlirStringCallback callback,
                                             void *userData) {
@@ -364,42 +344,6 @@ MLIR_CAPI_EXPORTED void beaverOperationDumpGeneric(MlirOperation op) {
   llvm::errs() << "\n";
 }
 
-template <typename T, typename EntityLookup, typename EntityGetter>
-T getIRDLDefinedEntity(MlirStringRef dialect, MlirStringRef name,
-                       MlirAttribute attrArr, EntityLookup lookup,
-                       EntityGetter getter) {
-  if (auto d =
-          unwrap(attrArr).getContext()->getOrLoadDialect(unwrap(dialect))) {
-    if (auto e = mlir::dyn_cast<ExtensibleDialect>(d)) {
-      if (auto definition = lookup(e, unwrap(name))) {
-        if (auto arr = mlir::dyn_cast<ArrayAttr>(unwrap(attrArr))) {
-          return getter(definition, arr.getValue());
-        }
-      }
-    }
-  }
-  return {};
-}
-
-MLIR_CAPI_EXPORTED MlirType beaverIRDLGetDefinedType(MlirStringRef dialect,
-                                                     MlirStringRef type,
-                                                     MlirAttribute params) {
-
-  return wrap(getIRDLDefinedEntity<Type>(
-      dialect, type, params,
-      [](auto d, auto name) { return d->lookupTypeDefinition(name); },
-      DynamicType::get));
-}
-
-MLIR_CAPI_EXPORTED MlirAttribute beaverIRDLGetDefinedAttr(
-    MlirStringRef dialect, MlirStringRef attr, MlirAttribute params) {
-
-  return wrap(getIRDLDefinedEntity<Attribute>(
-      dialect, attr, params,
-      [](auto d, auto name) { return d->lookupAttrDefinition(name); },
-      DynamicAttr::get));
-}
-
 MLIR_CAPI_EXPORTED MlirGreedyRewriteDriverConfig
 beaverGreedyRewriteDriverConfigGet() {
   return mlirGreedyRewriteDriverConfigCreate();
@@ -424,378 +368,4 @@ MLIR_CAPI_EXPORTED MlirType beaverDenseElementsAttrGetType(MlirAttribute attr) {
 
 MLIR_CAPI_EXPORTED intptr_t beaverShapedTypeGetNumElements(MlirType type) {
   return llvm::cast<ShapedType>(unwrap(type)).getNumElements();
-}
-
-#include "mlir-c/Debug.h"
-MLIR_CAPI_EXPORTED void beaverSetGlobalDebugTypes(const MlirStringRef *types,
-                                                  intptr_t n) {
-  // Convert MlirStringRef array to array of C strings
-  std::vector<const char *> cstrings;
-  cstrings.reserve(n);
-  for (intptr_t i = 0; i < n; ++i) {
-    cstrings.push_back(beaverStringRefGetData(types[i]));
-  }
-
-  // Call the underlying MLIR function
-  mlirSetGlobalDebugTypes(cstrings.data(), n);
-}
-
-#include "mlir/Dialect/GPU/IR/GPUDialect.h"
-
-MLIR_CAPI_EXPORTED MlirStringRef beaverGetNumWorkgroupAttributionsAttrName() {
-  return wrap(llvm::StringRef("workgroup_attributions"));
-}
-
-MLIR_CAPI_EXPORTED MlirStringRef beaverGetContainerModuleAttrName() {
-  return wrap(gpu::GPUDialect::getContainerModuleAttrName());
-}
-
-//===----------------------------------------------------------------------===//
-// Action Tracing
-//===----------------------------------------------------------------------===//
-
-using namespace mlir::tracing;
-
-namespace {
-
-/// Render a location as a compact string without full IR text.
-static std::string printLocation(Location loc) {
-  std::string buffer;
-  llvm::raw_string_ostream os(buffer);
-  loc.print(os);
-  os.flush();
-  return buffer;
-}
-
-/// Serialize one IR unit into a compact JSON object: {"kind": ..., "name": ...,
-/// "loc": ...}. Never materializes full IR text.
-static void serializeIRUnit(const IRUnit &unit, llvm::raw_string_ostream &os) {
-  os << "{\"kind\":";
-  if (isa<Operation *>(unit)) {
-    auto *op = cast<Operation *>(unit);
-    os << "\"operation\",\"name\":\"";
-    os.write_escaped(op->getName().getStringRef());
-    os << "\",\"loc\":\"";
-    os.write_escaped(printLocation(op->getLoc()));
-    os << "\"}";
-    return;
-  }
-  if (isa<Region *>(unit)) {
-    os << "\"region\"}";
-    return;
-  }
-  if (isa<Block *>(unit)) {
-    os << "\"block\"}";
-    return;
-  }
-  if (isa<Value>(unit)) {
-    Value value = cast<Value>(unit);
-    os << "\"value\",\"loc\":\"";
-    os.write_escaped(printLocation(value.getLoc()));
-    os << "\"}";
-    return;
-  }
-  os << "\"unknown\"}";
-}
-
-/// Serialize the IR units associated with an action.
-static std::string serializeIRUnits(ArrayRef<IRUnit> units) {
-  std::string buffer;
-  llvm::raw_string_ostream os(buffer);
-  os << "[";
-  bool first = true;
-  for (const IRUnit &unit : units) {
-    if (!first)
-      os << ",";
-    first = false;
-    serializeIRUnit(unit, os);
-  }
-  os << "]";
-  os.flush();
-  return buffer;
-}
-
-class ActionTracingSession {
-public:
-  ActionTracingSession(MLIRContext *context, std::string filterJson,
-                       std::string locationJson, std::string skipJson,
-                       std::string limitJson)
-      : context(context), filterTags(parseTagList(filterJson)),
-        filterLocations(parseStringList(locationJson)),
-        skipCounts(parseCountMap(skipJson)), limits(parseCountMap(limitJson)),
-        controlCallback([this](const ActionActiveStack *stack) {
-          return controlAction(stack);
-        }),
-        executionContext(controlCallback) {
-    // Breakpoints make the control callback run: ExecutionContext only invokes
-    // the callback when a breakpoint matches the action. Tags named in
-    // :skip/:limit therefore get a breakpoint so their first occurrences can be
-    // skipped or limited.
-    for (const auto &entry : skipCounts)
-      tagBreakpoints.addBreakpoint(entry.first);
-    for (const auto &entry : limits)
-      tagBreakpoints.addBreakpoint(entry.first);
-    executionContext.addBreakpointManager(&tagBreakpoints);
-  }
-
-  ~ActionTracingSession() {
-    context->registerActionHandler(nullptr);
-    // Drop any pending events; the session is being torn down.
-    std::lock_guard<std::mutex> lock(mutex);
-    events.clear();
-  }
-
-  void attach() {
-    context->registerActionHandler(
-        [this](function_ref<void()> transform, const Action &action) {
-          executionContext(transform, action);
-        });
-    executionContext.registerObserver(&observer);
-  }
-
-  bool drain(MlirBeaverActionEventsCallback callback, void *userData) {
-    std::vector<std::string> drained;
-    {
-      std::lock_guard<std::mutex> lock(mutex);
-      drained.swap(events);
-    }
-    std::string buffer;
-    llvm::raw_string_ostream os(buffer);
-    os << "[";
-    bool first = true;
-    for (const std::string &event : drained) {
-      if (!first)
-        os << ",";
-      first = false;
-      os << event;
-    }
-    os << "]";
-    os.flush();
-    callback(buffer.c_str(), userData);
-    return true;
-  }
-
-  MLIRContext *getContext() const { return context; }
-
-private:
-  struct Observer final : public ExecutionContext::Observer {
-    Observer(ActionTracingSession *session) : session(session) {}
-
-    ActionTracingSession *session;
-
-    void beforeExecute(const ActionActiveStack *stack, Breakpoint *,
-                       bool willExecute) override {
-      if (!willExecute)
-        return;
-      const Action &action = stack->getAction();
-      if (!session->shouldObserve(action))
-        return;
-      std::string description;
-      llvm::raw_string_ostream os(description);
-      action.print(os);
-      os.flush();
-      std::string event = "{\"phase\":\"before\",\"tag\":\"";
-      event += escapeJson(action.getTag());
-      event += "\",\"depth\":";
-      event += std::to_string(stack->getDepth());
-      event += ",\"description\":\"";
-      event += escapeJson(description);
-      event += "\",\"ir_units\":";
-      event += serializeIRUnits(action.getContextIRUnits());
-      event += ",\"t_ns\":";
-      event += std::to_string(nowNs());
-      event += "}";
-      session->push(event);
-    }
-
-    void afterExecute(const ActionActiveStack *stack) override {
-      const Action &action = stack->getAction();
-      if (!session->shouldObserve(action))
-        return;
-      std::string event = "{\"phase\":\"after\",\"tag\":\"";
-      event += escapeJson(action.getTag());
-      event += "\",\"depth\":";
-      event += std::to_string(stack->getDepth());
-      event += ",\"t_ns\":";
-      event += std::to_string(nowNs());
-      event += "}";
-      session->push(event);
-    }
-  };
-
-  ExecutionContext::Control controlAction(const ActionActiveStack *stack) {
-    const Action &action = stack->getAction();
-    std::string tag = action.getTag().str();
-    std::lock_guard<std::mutex> lock(mutex);
-
-    auto skipIt = skipCounts.find(tag);
-    if (skipIt != skipCounts.end()) {
-      if (skipIt->second > 0) {
-        skipIt->second--;
-        return ExecutionContext::Skip;
-      }
-      skipCounts.erase(skipIt);
-    }
-
-    auto limitIt = limits.find(tag);
-    if (limitIt != limits.end()) {
-      if (limitIt->second == 0)
-        return ExecutionContext::Skip;
-      limitIt->second--;
-    }
-
-    return ExecutionContext::Apply;
-  }
-
-  bool shouldObserve(const Action &action) const {
-    if (!filterTags.empty()) {
-      std::string tag = action.getTag().str();
-      if (std::find(filterTags.begin(), filterTags.end(), tag) ==
-          filterTags.end())
-        return false;
-    }
-    if (!filterLocations.empty()) {
-      bool found = false;
-      for (const IRUnit &unit : action.getContextIRUnits()) {
-        std::optional<Location> loc;
-        if (isa<Operation *>(unit))
-          loc = cast<Operation *>(unit)->getLoc();
-        else if (isa<Value>(unit))
-          loc = cast<Value>(unit).getLoc();
-        if (!loc)
-          continue;
-        std::string printed = printLocation(*loc);
-        for (const std::string &needle : filterLocations) {
-          if (printed.find(needle) != std::string::npos) {
-            found = true;
-            break;
-          }
-        }
-        if (found)
-          break;
-      }
-      if (!found)
-        return false;
-    }
-    return true;
-  }
-
-  void push(std::string event) {
-    std::lock_guard<std::mutex> lock(mutex);
-    events.push_back(std::move(event));
-  }
-
-  static std::vector<std::string> parseTagList(llvm::StringRef json) {
-    return parseStringList(json);
-  }
-
-  static std::vector<std::string> parseStringList(llvm::StringRef json) {
-    if (json.empty())
-      return {};
-    std::vector<std::string> tags;
-    // Extremely small JSON array parser: ["a","b"]
-    llvm::StringRef rest = json.trim();
-    rest = rest.drop_front().drop_back(); // strip [ ]
-    while (!rest.empty()) {
-      rest = rest.trim().ltrim(',');
-      if (rest.empty())
-        break;
-      rest = rest.trim();
-      if (!rest.consume_front("\""))
-        break;
-      size_t end = rest.find('"');
-      if (end == llvm::StringRef::npos)
-        break;
-      tags.push_back(rest.substr(0, end).str());
-      rest = rest.substr(end + 1);
-    }
-    return tags;
-  }
-
-  static std::unordered_map<std::string, uint64_t>
-  parseCountMap(llvm::StringRef json) {
-    std::unordered_map<std::string, uint64_t> result;
-    if (json.empty())
-      return result;
-    llvm::StringRef rest = json.trim().drop_front().drop_back(); // strip { }
-    while (!rest.empty()) {
-      rest = rest.trim().ltrim(',');
-      if (rest.empty())
-        break;
-      rest = rest.trim();
-      if (!rest.consume_front("\""))
-        break;
-      size_t end = rest.find('"');
-      if (end == llvm::StringRef::npos)
-        break;
-      std::string key = rest.substr(0, end).str();
-      rest = rest.substr(end + 1).trim();
-      if (!rest.consume_front(":"))
-        break;
-      rest = rest.trim();
-      size_t numEnd = rest.find_first_of(",}");
-      uint64_t value = 0;
-      rest.substr(0, numEnd).getAsInteger(10, value);
-      result[std::move(key)] = value;
-      rest = rest.substr(numEnd == llvm::StringRef::npos ? rest.size() : numEnd);
-    }
-    return result;
-  }
-
-  static std::string escapeJson(llvm::StringRef value) {
-    std::string out;
-    llvm::raw_string_ostream os(out);
-    os.write_escaped(value);
-    os.flush();
-    return out;
-  }
-
-  static uint64_t nowNs() {
-    return std::chrono::duration_cast<std::chrono::nanoseconds>(
-               std::chrono::steady_clock::now().time_since_epoch())
-        .count();
-  }
-
-  MLIRContext *context;
-  std::vector<std::string> filterTags;
-  std::vector<std::string> filterLocations;
-  std::unordered_map<std::string, uint64_t> skipCounts;
-  std::unordered_map<std::string, uint64_t> limits;
-  std::function<ExecutionContext::Control(const ActionActiveStack *)>
-      controlCallback;
-  tracing::TagBreakpointManager tagBreakpoints;
-  ExecutionContext executionContext;
-  Observer observer{this};
-  std::mutex mutex;
-  std::vector<std::string> events;
-};
-
-} // namespace
-
-MLIR_CAPI_EXPORTED MlirBeaverActionTracing
-mlirBeaverActionTracingAttach(MlirContext context, MlirStringRef filter_json,
-                              MlirStringRef location_json,
-                              MlirStringRef skip_json, MlirStringRef limit_json) {
-  auto *session = new ActionTracingSession(
-      unwrap(context),
-      llvm::StringRef(filter_json.data, filter_json.length).str(),
-      llvm::StringRef(location_json.data, location_json.length).str(),
-      llvm::StringRef(skip_json.data, skip_json.length).str(),
-      llvm::StringRef(limit_json.data, limit_json.length).str());
-  session->attach();
-  return MlirBeaverActionTracing{session};
-}
-
-MLIR_CAPI_EXPORTED bool mlirBeaverActionTracingDrain(
-    MlirBeaverActionTracing tracing, MlirBeaverActionEventsCallback callback,
-    void *user_data) {
-  auto *session = static_cast<ActionTracingSession *>(tracing.ptr);
-  if (!session)
-    return false;
-  return session->drain(callback, user_data);
-}
-
-MLIR_CAPI_EXPORTED void
-mlirBeaverActionTracingDetach(MlirBeaverActionTracing tracing) {
-  delete static_cast<ActionTracingSession *>(tracing.ptr);
 }
